+++ src/Request/Checkpoint.php
@@ -72,10 +72,11 @@ class Checkpoint extends RequestCollection
 
         return $this->ig->request($checkpointUrl)
            ->setNeedsAuth(false)
+           ->setSignedPost(false)
            ->addPost('choice', $verificationMethod)
            ->addPost('guid', $this->ig->uuid)
            ->addPost('device_id', $this->ig->device_id)
-           ->addPost('_csrftoken', $this->ig->client->getToken())
+           //->addPost('_csrftoken', $this->ig->client->getToken())
            ->getResponse(new Response\CheckpointResponse());
     }
 
@@ -95,10 +96,11 @@ class Checkpoint extends RequestCollection
     {
         return $this->ig->request($checkpointUrl)
             ->setNeedsAuth(false)
+            ->setSignedPost(false)
             ->addPost('security_code', $verificationCode)
             ->addPost('guid', $this->ig->uuid)
             ->addPost('device_id', $this->ig->device_id)
-            ->addPost('_csrftoken', $this->ig->client->getToken())
+            //->addPost('_csrftoken', $this->ig->client->getToken())
             ->getResponse(new Response\LoginResponse());
     }
 
@@ -121,10 +123,11 @@ class Checkpoint extends RequestCollection
         $afterLogin = false)
     {
         $request = $this->ig->request($checkpointUrl)
+            ->setSignedPost(false)
             ->addPost('phone_number', $phoneNumber)
             ->addPost('guid', $this->ig->uuid)
-            ->addPost('device_id', $this->ig->device_id)
-            ->addPost('_csrftoken', $this->ig->client->getToken());
+            ->addPost('device_id', $this->ig->device_id);
+            //->addPost('_csrftoken', $this->ig->client->getToken());
         if ($afterLogin === false) {
             $request->setNeedsAuth(false);
         } else {
@@ -154,10 +157,11 @@ class Checkpoint extends RequestCollection
         $afterLogin = false)
     {
         $request = $this->ig->request($checkpointUrl)
+            ->setSignedPost(false)
             ->addPost('email', $email)
             ->addPost('guid', $this->ig->uuid)
-            ->addPost('device_id', $this->ig->device_id)
-            ->addPost('_csrftoken', $this->ig->client->getToken());
+            ->addPost('device_id', $this->ig->device_id);
+            //->addPost('_csrftoken', $this->ig->client->getToken());
         if ($afterLogin === false) {
             $request->setNeedsAuth(false);
         } else {